backend/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
python-dotenv
requests
pandas
numpy
matplotlib
scikit-learn
xgboost